pyproject.toml
[tool.poetry]
name = "smart-lang-chain"
version = "0.1.0"
description = ""
authors = ["Andrii Peliak <andrii.peliak@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
langchain = ">=0.3.26,<0.4.0"
langchain-openai = ">=0.3.25,<0.4.0"
python-dotenv = ">=1.1.1,<2.0.0"
langchain-community = "^0.3.26"
transformers = "^4.52.4"
torch = "^2.7.1"
langchain-huggingface = "^0.3.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
